tutor/requirements.txt
# TUTOR 项目依赖
# Python 3.9+

# 核心依赖
requests>=2.28.0
pyyaml>=6.0

# CLI
typer>=0.9.0
rich>=13.0.0

# API
fastapi>=0.100.0  # MVP延后实现Web
uvicorn>=0.23.0   # MVP延后

# 测试
pytest>=7.0.0
pytest-mock>=3.10.0

# 数据库
sqlite3  # Python标准库

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0
pydantic[email]>=2.0.0
